refactor#51 임포트 수정 notification 연관관계 매핑 수정
