--- omc/src/main/java/com/hotsix/omc/notification/type/PageType.java
@@ -1,5 +1,0 @@
-package com.hotsix.omc.notification.type;
-
-public enum PageType {
-    CAR_INSPECTION, REVIEW
-}
