+++ omc/src/main/java/com/hotsix/omc/s3/dto/FileDetailDto.java
@@ -1,6 +1,5 @@
 package com.hotsix.omc.s3.dto;
 
-import com.hotsix.omc.s3.util.MultipartUtil;
 import lombok.*;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -23,13 +22,13 @@ public class FileDetailDto {
     private LocalDateTime createdAt = LocalDateTime.now();
 
     public static FileDetailDto multipartOf(MultipartFile multipartFile) {
-        final String fileId = MultipartUtil.createFileId();
-        final String format = MultipartUtil.getFormat(multipartFile.getContentType());
+        final String fileId = com.hotsix.omc.util.MultipartUtil.createFileId();
+        final String format = com.hotsix.omc.util.MultipartUtil.getFormat(multipartFile.getContentType());
         return FileDetailDto.builder()
                 .id(fileId)
                 .name(multipartFile.getOriginalFilename())
                 .format(format)
-                .path(MultipartUtil.createPath(fileId, format))
+                .path(com.hotsix.omc.util.MultipartUtil.createPath(fileId, format))
                 .bytes(multipartFile.getSize())
                 .build();
     }
